+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - profiler now correctly removes `::` separators from function names which were monomorphised with different types
 
+### Changed
+
+- upgraded cost map (versioned constants) to starknet 0.14.1
+
 ## [0.10.0] - 2025-07-18
 
 ### Changed